+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- `EvalState::store()`. Consumers need to track the `Store` themselves.
+
 ## [0.2.1] - 2026-05-16
 
 ### Added